requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
ecdsa==0.18.0
idna==3.4
pyasn1==0.4.8
python-decouple==3.6
python-jose==3.3.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
